wp31_services/energy_analyzer/requirements.txt
grpcio>=1.59.0
protobuf>=4.25.0
grpcio-health-checking>=1.59.0
pandas